add console hook and settings
