+++ napari/utils/settings/_defaults.py
@@ -6,7 +6,8 @@ from typing import List, Tuple
 
 from pydantic import BaseSettings, Field
 
-from napari.utils.events.evented_model import EventedModel
+from ..events.evented_model import EventedModel
+from ..notifications import NotificationSeverity
 
 
 class QtBindingChoice(str, Enum):
@@ -56,6 +57,10 @@ class ApplicationSettings(BaseSettings, EventedModel):
     window_state: str = None
     window_statusbar: bool = True
     preferences_size: Tuple[int, int] = None
+    gui_notification_level: NotificationSeverity = NotificationSeverity.INFO
+    console_notification_level: NotificationSeverity = (
+        NotificationSeverity.NONE
+    )
 
     class Config:
         # Pydantic specific configuration
