+++ napari/utils/settings/_manager.py
@@ -8,7 +8,7 @@ from appdirs import user_config_dir
 from pydantic import ValidationError
 from yaml import safe_dump, safe_load
 
-from ..settings import CORE_SETTINGS
+from ._defaults import CORE_SETTINGS, ApplicationSettings, PluginSettings
 
 
 class SettingsManager:
@@ -47,6 +47,8 @@ class SettingsManager:
     _FILENAME = "settings.yaml"
     _APPNAME = "Napari"
     _APPAUTHOR = "Napari"
+    application: ApplicationSettings
+    plugin: PluginSettings
 
     def __init__(self, config_path: str = None, save_to_disk: bool = True):
         self._config_path = (
